+++ ChangeLog.md
@@ -1,5 +1,13 @@
 # Revision history for reflex-vty
 
+## Unreleased
+
+* Re-design `textInput`, `TextInput` and `TextInputConfig`.
+  * Allows users to implement more complex behavior.
+  * `_textInputConfig_modify` is now applied to the text-value of `textInput`
+    after user input events such as mouse clicks and keyboard input.
+    This may change the observable behavior.
+
 ## 0.2.0.1
 
 * Loosen version bounds on ref-tf and vty